fix table not found usage
